requirements.txt
discord.py
mariadb
bcrypt
python-dotenv
uuid
secrets